Do not call it 1.0 yet
module has been only used in one environment within one use case.  The
API is still not organized well (there might be some bad decisions),
documentation needs some reviewing, ant the name of the module is likely
to change in forseeable future.

Therefore it seems too brave to call it 1.0, I'd be happier if the above
was finished before that.

Also, I believe it's a good idea to adhere to Semantic Versioning model,
therefore 0.1.0.